requirements.txt
# FastAPI and Server
fastapi==0.109.1
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Supabase
supabase==2.3.4

# Authentication & Security
bcrypt==4.1.2
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP Client
httpx>=0.24,<0.26

# Logging
python-json-logger==2.0.7

# Redis for rate limiting
redis==5.0.1

# Email validation
email-validator==2.3.0

# SMTP for email sending
aiosmtplib==3.0.1

# Monitoring and Observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.40.3
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-httpx==0.43b0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
